se añaden modulos nuevos: loadAssets: carga modelos y scalers data: carga y transforma los datos main: para llamar a la ejecución del modelo
