+++ apollo/src/loadAssets.py
@@ -1,9 +1,11 @@
+import sys
 import pickle
 import logging
 import pandas as pd
 from getData.extract import get_forex
 from processData.processing import setup_data, get_indicators
-
+#sys.path.append('./src/assets/')
+#sys.path.append('./src')
 
 class Assets():
 
@@ -13,14 +15,13 @@ class Assets():
 
     def load_vals(self):
         # variables para High
-        logging.info('************* Cargando Variables **************')
         variablesh = pd.read_csv(
-            f'./src/assets/{self.model_version}/variablesHigh.csv')
+            f'./src/assets/{self.model_version}/variables/variablesHigh.csv')
         variablesh = list(variablesh['0'].values)
 
         # variables para Low
         variablesl = pd.read_csv(
-            f'./src/assets/{self.model_version}/variablesLow.csv')
+            f'src/assets/{self.model_version}/variables/variablesLow.csv')
         variablesl = list(variablesl['0'].values)
 
         return variablesh, variablesl
@@ -29,7 +30,7 @@ class Assets():
     def load_models(self):
         files = []
         logging.info(
-            f'************* Cargando Modelos {self.model_version}Gradient Boosting (High)**************')
+            f'************* Cargando Modelos ({self.model_version}) Gradient Boosting (High)**************')
 
         files.append(pickle.load(
             open(f'./src/assets/{self.model_version}/gbHigh-6.h5', 'rb')))
@@ -59,7 +60,7 @@ class Assets():
             open(f'./src/assets/{self.model_version}/gbHigh6.h5', 'rb')))
 
         logging.info(
-            '************* Cargando Modelos  Gradient Boosting (Low)**************')
+            f'************* Cargando Modelos ({self.model_version}) Gradient Boosting (Low)**************')
         files.append(pickle.load(
             open(f'./src/assets/{self.model_version}/gbLow-6.h5', 'rb')))
         files.append(pickle.load(
@@ -86,39 +87,6 @@ class Assets():
             open(f'./src/assets/{self.model_version}/gbLow5.h5', 'rb')))
         files.append(pickle.load(
             open(f'./src/assets/{self.model_version}/gbLow6.h5', 'rb')))
-
-        models = {}
-        
-        models = {f'Xh_gbH-{i}': file_model.predict_proba(Xh) for i,file_model in enumerate(files[:len(files)/2])}
-
-        
-        models['Xh_gbH-6'] = gbH_6.predict_proba(Xh)[:, 1]
-        models['Xh_gbH-5'] = gbH_5.predict_proba(Xh)[:, 1]
-        models['Xh_gbH-4'] = gbH_4.predict_proba(Xh)[:, 1]
-        models['Xh_gbH-3'] = gbH_3.predict_proba(Xh)[:, 1]
-        models['Xh_gbH-2'] = gbH_2.predict_proba(Xh)[:, 1]
-        models['Xh_gbH-1'] = gbH_1.predict_proba(Xh)[:, 1]
-        models['Xh_gbH0'] = gbH0.predict_proba(Xh)[:, 1]
-        models['Xh_gbH1'] = gbH1.predict_proba(Xh)[:, 1]
-        models['Xh_gbH2'] = gbH2.predict_proba(Xh)[:, 1]
-        models['Xh_gbH3'] = gbH3.predict_proba(Xh)[:, 1]
-        models['Xh_gbH4'] = gbH4.predict_proba(Xh)[:, 1]
-        models['Xh_gbH5'] = gbH5.predict_proba(Xh)[:, 1]
-        models['Xh_gbH6'] = gbH6.predict_proba(Xh)[:, 1]
-
-        models['Xl_gbl-6'] = gbL_6.predict_proba(Xl)[:, 1]
-        models['Xl_gbl-5'] = gbL_5.predict_proba(Xl)[:, 1]
-        models['Xl_gbl-4'] = gbL_4.predict_proba(Xl)[:, 1]
-        models['Xl_gbl-3'] = gbL_3.predict_proba(Xl)[:, 1]
-        models['Xl_gbl-2'] = gbL_2.predict_proba(Xl)[:, 1]
-        models['Xl_gbl-1'] = gbL_1.predict_proba(Xl)[:, 1]
-        models['Xl_gbl0'] = gbL0.predict_proba(Xl)[:, 1]
-        models['Xl_gbl1'] = gbL1.predict_proba(Xl)[:, 1]
-        models['Xl_gbl2'] = gbL2.predict_proba(Xl)[:, 1]
-        models['Xl_gbl3'] = gbL3.predict_proba(Xl)[:, 1]
-        models['Xl_gbl4'] = gbL4.predict_proba(Xl)[:, 1]
-        models['Xl_gbl5'] = gbL5.predict_proba(Xl)[:, 1]
-        models['Xl_gbl6'] = gbL6.predict_proba(Xl)[:, 1]
 
         return files
